requirements.txt
flask
flask-cors
pillow
numpy
tensorflow-cpu
scipy
gunicorn
